requirements.txt
streamlit
pandas
pandasql==0.7.3
python-dotenv
groq
semantic-router[local]
chromadb
sentence_transformers